[Binding][iCloud] Package org.openhab.binding.icloud.internal.json.icloud renamed to org.openhab.binding.icloud.internal.json
