+++ addons/binding/org.openhab.binding.icloud/src/main/java/org/openhab/binding/icloud/handler/ICloudAccountBridgeHandler.java
@@ -32,8 +32,8 @@ import org.openhab.binding.icloud.internal.Connection;
 import org.openhab.binding.icloud.internal.DeviceInformationParser;
 import org.openhab.binding.icloud.internal.configuration.AccountThingConfiguration;
 import org.openhab.binding.icloud.internal.discovery.DeviceDiscovery;
-import org.openhab.binding.icloud.internal.json.icloud.DeviceInformation;
-import org.openhab.binding.icloud.internal.json.icloud.JSONRootObject;
+import org.openhab.binding.icloud.internal.json.DeviceInformation;
+import org.openhab.binding.icloud.internal.json.JSONRootObject;
 import org.osgi.framework.ServiceRegistration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
