+++ addons/binding/org.openhab.binding.icloud/src/main/java/org/openhab/binding/icloud/internal/discovery/DeviceDiscovery.java
@@ -17,7 +17,7 @@ import org.eclipse.smarthome.config.discovery.DiscoveryResult;
 import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.openhab.binding.icloud.handler.ICloudAccountBridgeHandler;
-import org.openhab.binding.icloud.internal.json.icloud.DeviceInformation;
+import org.openhab.binding.icloud.internal.json.DeviceInformation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
