+++ addons/binding/org.openhab.binding.icloud/src/main/java/org/openhab/binding/icloud/internal/json/Timezone.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.icloud.internal.json.icloud;
+package org.openhab.binding.icloud.internal.json;
 
 /**
  *
